+++ package.json
@@ -119,7 +119,7 @@
     "imagemin-webpack-plugin": "2.4.2",
     "is-ipfs": "5.0.0",
     "jest-skipped-reporter": "0.0.5",
-    "marked": "2.0.3",
+    "marked": "2.0.7",
     "multicoin-address-validator": "0.4.5",
     "node-fetch": "2.6.1",
     "node-sass": "6.0.0",
