+++ src/modules/access-wallet/common/index.js
@@ -7,9 +7,9 @@ import {
   KeepkeyWallet,
   BCVaultWallet,
   CoolWallet
-} from '@/modules/access-wallet/hardware/handlers/handlerAccessWalletHardware';
+} from '@/modules/access-wallet/hardware/handlers';
 import WalletInterface from './WalletInterface';
-import { MnemonicWallet } from './software';
+import { MnemonicWallet } from '../software/handlers';
 import {
   MewConnectWallet,
   WalletConnectWallet,
