+++ src/modules/access-wallet/common/utils.js
