+++ src/modules/access-wallet/common/walletConfigs.js
