+++ src/modules/access-wallet/hardware/handlers/hardwares/bcvault/index.js
@@ -2,7 +2,7 @@ import { Transaction } from 'ethereumjs-tx';
 import { bufferToHex } from 'ethereumjs-util';
 import * as bc from 'bc-vault-js';
 import { WALLET_TYPES } from '../../configs/configWalletTypes';
-import HDWalletInterface from '@/modules/wallets/utils/HDWalletInterface.js';
+import HDWalletInterface from '@/modules/access-wallet/common/HDWalletInterface';
 import errorHandler from './errorHandler';
 import BigNumber from 'bignumber.js';
 import commonGenerator from '@/core/helpers/commonGenerator';
@@ -14,7 +14,7 @@ import {
   sanitizeHex,
   getSignTransactionObject,
   calculateChainIdFromV
-} from '@/modules/wallets/utils/utils.js';
+} from '@/modules/access-wallet/common/utils';
 
 const NEED_PASSWORD = false;
