+++ src/modules/access-wallet/hardware/handlers/hardwares/bitbox/index.js
@@ -4,7 +4,7 @@ import DigitalBitboxUsb from './digitalBitboxUsb';
 import DigitalBitboxEth from './digitalBitboxEth';
 import { WALLET_TYPES } from '../../configs/configWalletTypes';
 import bip44Paths from '@/modules/access-wallet/hardware/handlers/bip44';
-import HDWalletInterface from '@/modules/wallets/utils/HDWalletInterface.js';
+import HDWalletInterface from '@/modules/access-wallet/common/HDWalletInterface.js';
 import { Toast, SENTRY } from '@/modules/toast/handler/handlerToast';
 import toBuffer from '@/core/helpers/toBuffer';
 import errorHandler from './errorHandler';
