+++ src/modules/access-wallet/hardware/handlers/hardwares/bitbox02/index.js
@@ -2,7 +2,7 @@ import { BitBox02API, getDevicePath, constants } from 'bitbox02-api';
 
 import { WALLET_TYPES } from '../../configs/configWalletTypes';
 import bip44Paths from '@/modules/access-wallet/hardware/handlers/bip44';
-import HDWalletInterface from '@/modules/wallets/utils/HDWalletInterface.js';
+import HDWalletInterface from '@/modules/access-wallet/common/HDWalletInterface.js';
 import * as HDKey from 'hdkey';
 import { Transaction } from 'ethereumjs-tx';
 import {
