+++ src/modules/access-wallet/hybrid/handlers/MEWconnect/errorHandler.js
