+++ src/modules/access-wallet/hybrid/handlers/WalletConnect/errorHandler.js
