+++ src/modules/access-wallet/hybrid/handlers/WalletConnect/index.js
