+++ src/modules/access-wallet/software/handlers/handlerAccessWalletSoftware.js
@@ -1,7 +1,7 @@
-import { MnemonicWallet } from '@/modules/wallets/utils/software';
-import WalletInterface from '@/modules/wallets/utils/WalletInterface';
+import { MnemonicWallet } from '@/modules/access-wallet/software/handlers';
+import WalletInterface from '@/modules/access-wallet/common/WalletInterface';
 import { WALLET_TYPES } from '@/modules/access-wallet/hardware/handlers/configs/configWalletTypes';
-import { unlockKeystore } from '@/modules/wallets/utils/helpers.js';
+import { unlockKeystore } from '@/modules/access-wallet/common/helpers';
 import errorHandler from './errorHandler';
 
 export default class AccessWallet {
