+++ src/modules/access-wallet/software/handlers/mnemonic/errorHandler.js
