+++ src/modules/confirmation/ModuleConfirmation.vue
@@ -101,7 +101,7 @@ import parseTokenData from '@/core/helpers/parseTokenData';
 import { EventBus } from '@/core/plugins/eventBus';
 import { setEvents } from '@/utils/web3-provider/methods/utils.js';
 import * as locStore from 'store';
-import { sanitizeHex } from '@/modules/wallets/utils/utils.js';
+import { sanitizeHex } from '@/modules/access-wallet/common/utils';
 export default {
   name: 'ConfirmationContainer',
   components: {
